Check PIL support for tostring and fromstring
The current code checks whether the methods tostring and fromstring
exist in a PIL image object like this:

  getattr(image, 'tobytes', image.tostring)

This code raises an exception when image doesn't have the tostring
method. That seems to be the case in recent pillow versions!

Check if the tostring or tobytes method exists (in this order).

This is related to #644.